iox: handle nil File/BufferFile on Close
It is valid to call Close on a nil *os.File.
To make swapping out types as easy as possible,
support the same behavior on *iox.File and
*iox.BufferFile.

+++ filer.go
@@ -224,6 +224,9 @@ func (file *File) remove() {
 
 // Close closes the underlying file descriptor and informs the Filer.
 func (file *File) Close() error {
+	if file == nil || file.File == nil {
+		return os.ErrInvalid
+	}
 	err := file.File.Close()
 	file.remove()
